Suggested privately that there's no need from fromIntegral here.
should be fromInteger, although because of the Num instance we
can just write the number directly.

+++ examples/SimpleConnection.hs
@@ -5,7 +5,7 @@ main = do
     ctx <- initConnectionContext
     con <- connectTo ctx $ ConnectionParams
                               { connectionHostname  = "www.example.com"
-                              , connectionPort      = fromIntegral 4567
+                              , connectionPort      = 4567
                               , connectionUseSecure = Nothing
                               , connectionUseSocks  = Nothing
                               }
