+++ examples/SocksConnection.hs
@@ -5,9 +5,9 @@ main = do
     ctx <- initConnectionContext
     con <- connectTo ctx $ ConnectionParams
                               { connectionHostname  = "www.example.com"
-                              , connectionPort      = fromIntegral 4567
+                              , connectionPort      = 4567
                               , connectionUseSecure = Nothing
-                              , connectionUseSocks  = Just $ SockSettingsSimple "localhost" (fromIntegral 1080)
+                              , connectionUseSocks  = Just $ SockSettingsSimple "localhost" 1080
                               }
     connectionPut con (B.singleton 0xa)
     r <- connectionGet con 1024
